+++ pyamg/relaxation/smoothing.py
@@ -434,9 +434,9 @@ def rho_block_D_inv_A(A, Dinv):
             raise ValueError('Dinv and A have incompatible dimensions')
 
         Dinv = sparse.bsr_array((Dinv,
-                                  np.arange(Dinv.shape[0]),
-                                  np.arange(Dinv.shape[0]+1)),
-                                 shape=A.shape)
+                                 np.arange(Dinv.shape[0], dtype=np.int32),
+                                 np.arange(Dinv.shape[0] + 1, dtype=np.int32)),
+                                shape=A.shape)
 
         # Don't explicitly form Dinv @ A
         def matvec(x):
